+++ example_channelfit.py
@@ -20,13 +20,14 @@ chan.makegrid(cubefits=cubefits, center=center, pa=pa, incl=incl,
               vsys=vsys, dist=dist, sigma=sigma, rmax=rmax, vlim=vlim)
 chan.fitting(Mstar_range=[0.1, 2.0],
              Rc_range=[30, 300],
-             cs_range=[0, 0.5],
-             fixed_params={'h1': 0, 'h2': -1, 'Rin': 0,
-                           'pI': 0, 'Ienv': 0,
-                           'xoff': 0, 'yoff': 0, 'voff': 0, 'incl': 0},
+             pI_range=[0, 2],
+             fixed_params={'cs': 0.4, 'h1': 0, 'h2': -1,
+                           'Rin': 0, 'Ienv': 0,
+                           'xoff': 0, 'yoff': 0, 'voff': 0,
+                           'incloff': 0, 'paoff':0},
              kwargs_emcee_corner={'nwalkers_per_ndim': 4,
-                                  'nburnin': 100,
-                                  'nsteps': 200,
+                                  'nburnin': 10,
+                                  'nsteps': 20,
                                   'rangelevel': 0.99},
              filename=filehead)
 p = chan.popt
